* Remove a slightly-hidden `toView . fromView` from the implementation
  of `>>=` for `SeqT`.

* Make `fmap` for `SeqT` do the most natural thing, rather than
  restructuring the `SeqT`. Whether this is better or worse will
  depend on the exact situation, but my bet is it's usually better,
  and it's more what one expects from `fmap`.

* Add a few more method implementations that might or might not matter.

* Improve `fmap` for `MSeq`.