@dagit 2021-07-23
Have you looked at the `inline` branch? That's where I've been exploring optimizations. I'm not sure which parts of that stuff I want to keep ultimately so that's why I'm doing it in a branch.

At one point I had a rewrite rule like this (and I don't think I meant to delete it):

```haskell
{-# RULES "toView/fromView" forall m. fromView (toView m) = m; #-}
```

I didn't check the rewrite summary, but I was hoping it would catch cases like the one you optimized away.

Another thing you can do is provide rewrites that are type specific like this:
```haskell
instance Monad m => Monad (SeqT m) where
  {-# INLINE CONLIKE return #-}
  {- NOINLINE [1] (>>=) #-}
  return = fromView . single
  (>>=) = seqBind
#if !MIN_VERSION_base(4,13,0)
  {-# INLINEABLE fail #-}
  fail = Fail.fail
#endif

seqBind :: Monad m => SeqT m a -> (a -> SeqT m b) -> SeqT m b
seqBind a b = go a b
  where
  go m f = fromView (toView m >>= \x -> case x of
    Empty -> return Empty
    h :< t -> toView (f h `seqAlt` (t `go` f)))
{-# INLINEABLE [2] seqBind #-}

seqBindIdentity :: Seq a -> (a -> Seq b) -> Seq b
seqBindIdentity a b = go a b
  where
  go m f = case toViewIdentity (runSeqT m) of
    Empty -> fromView (Identity Empty)
    h :< t -> f h `seqAltIdentity` (t `go` f)
{-# INLINEABLE[2] seqBindIdentity #-}

{-# RULES
      "seqBindIdentity" [3] seqBind = seqBindIdentity;
  #-}
```

So that does a few things. It splits `seqBind` into a worker wrapper because GHC tends to be better at optimizing the result (it can safely inline `seqBind` because the recursion was moved to the where-clause): https://wiki.haskell.org/Worker_wrapper

And I was experimenting with optimizing specifically for the `SeqT Identity` case, although I'm not sure that actually does much in practice. I think maybe `Identity` is already optimized through other GHC tricks? Anyway, the trick I want to highlight is the rewrite `seqBind = seqBindIdentity`. It only applies when the types match on both sides. So in the case where GHC knows statically that you are working with `Seq` it can use the simplified version of bind. It doesn't seem to add much in this specific usage, but I feel that the trick overall is worth mentioning.

And finally, I started adding `CONLIKE` to all the functions that are meant to be cheap. This tells GHC that it's okay to duplicate that work if it leads to more `RULES` firing. As is the case for simple "constructor-like" functions.

I think there should be some fusion rewrites that we'd benefit. And any sort of identities to simplify `>>=` and `<|>` are going to be a big win.