@treeowl 2021-07-23
No, I was not aware of that branch. We definitely *don't* want to rely on rewrite rules to avoid unnecessary `toView . fromView` in `>>=`. That just depends too much on how the inlining goes and other details of the optimization process.

I would *expect* that GHC's specialization would do a pretty good job with `Identity`. You can certainly double check, but I wouldn't expect that to have much impact.