@treeowl 2021-07-23
I don't *think* you're being careful enough with `CONLIKE`. `single`, for example, and all the things that use it, call `pure` in the underlying monad. That's often cheap, but sometimes not. Consider, for example, the `Monad` instance for size-indexed vectors. If you duplicate `pure`, you could allocate multiple 10,000 element vectors when there's only supposed to be one.