@dagit 2021-07-23
Like I said, not all of this stuff should end up in the public API. I wanted to try a lot of different things and see what helps.