@treeowl 2021-07-23
I think it probably makes sense to start with this pull request, which I believe is pretty solid. The part of your `inline` branch that I think is most promising is the static argument stuff (what you called worker-wrapper). I'm not convinced by most of the rest. The `toView . fromView` and `fromView . toView` strike me as desirable, but also slightly dangerous—if the underlying `Monad` instance is unlawful, then optimization flags could affect program behavior.