@treeowl 2021-07-23
Question: were you able to measure an improvement from the static argument transformation? I can't immediately tell whether that gives enough optimization opportunities in this particular context.