@dagit 2021-07-23
> 
> 
> Question: were you able to measure an improvement from the static argument transformation? I can't immediately tell whether that gives enough optimization opportunities in this particular context.

I haven't spent any time yet trying to figure out specifically which changes lead to which time improvement. I've actually been thinking about the problem differently: How much can we close the gap? For me it's useful to have a performance goal in mind and that's what I've been focusing on figuring out.

That said, I don't see any reason not to do that transformation in this case and with more principled application of worker/wrapper transformation there might be a versions of `>>=`/`<|>` that works on a more "unwrapped" version of the monad analogous to worker/wrapper allowing you to computing factorial on unboxed ints internally.

I figured once I had a bunch of candidate optimizations I could look at their contribution individually and when they're safe and all that stuff.

I do know that improvements to `>>=` and `<|>` tend to give the biggest win, but `toView/fromView` are also very important.

I'm currently wondering if deforestation can be applied in some cases. Imagine computations where values are consumed without ever doing monadic reflection. Is there any hope of removing the intermediate queues in that case? I've never actually looked into deforestation so I hope to read through some papers this weekend to get a sense of whether the ideas would work at all. I realize they're sort of the opposite of what reflection without remorse is trying to achieve, but my hope is that something like deforestation or unwrapping could apply in special cases where inspection is not happening.