@treeowl 2021-07-23
The sort of fusion framework you're suggesting seems to have gone somewhat out of fashion, mostly because it's painfully finicky. `vector` uses a "bundle" framework to do this sort of fuse this/don't fuse that kind of thing, but it's *hideously* complicated and seems somewhat brittle.