@dagit 2021-07-23
> I think it probably makes sense to start with this pull request, which I believe is pretty solid. 

Yeah I agree. This one and the instances PR both look good and I'll merge them this evening. I'd do it right now except I kind of want to run my benchmark on this one before/after so I can see the delta on it.