@dagit 2021-07-23
I was going to read this document first which appears to cover the development and history of it: http://www.ccs.neu.edu/home/amal/course/7480-s12/deforestation-notes.pdf

But if there's a specific technique or paper you think I should look at, let me know.