@dagit 2021-07-24
Okay, so I just ran my super secret benchmark (sorry, I really need to upload it, but we do have benchmarks now thanks to another contributor), and some how these changes slow things down and I haven't quite figured out how yet but I can share the timing difference:

Before:
```
SeqT IO with msplit:
Batch generated in: 0.545504 seconds
Batch generated in: 0.52716 seconds
Batch generated in: 0.511638 seconds
Batch generated in: 0.540144 seconds
Batch generated in: 0.529143 seconds
Total time: 2.653746 seconds

SeqT Identity with msplit:
Batch generated in: 0.606576 seconds
Batch generated in: 0.626024 seconds
Batch generated in: 0.612829 seconds
Batch generated in: 0.5735899999999999 seconds
Batch generated in: 0.606266 seconds
Total time: 3.025437 seconds

SeqT Identity without msplit:
Batch generated in: 0.615291 seconds
Batch generated in: 0.613065 seconds
Batch generated in: 0.6118 seconds
Batch generated in: 0.568402 seconds
Batch generated in: 0.635086 seconds
Total time: 3.043796 seconds
```

After these changes:
```
SeqT IO with msplit:
Batch generated in: 0.702476 seconds
Batch generated in: 0.678819 seconds
Batch generated in: 0.663562 seconds
Batch generated in: 0.6896289999999999 seconds
Batch generated in: 0.7065319999999999 seconds
Total time: 3.441195 seconds

SeqT Identity with msplit:
Batch generated in: 0.658728 seconds
Batch generated in: 0.7158289999999999 seconds
Batch generated in: 0.682816 seconds
Batch generated in: 0.6083649999999999 seconds
Batch generated in: 0.687726 seconds
Total time: 3.353606 seconds

SeqT Identity without msplit:
Batch generated in: 0.634589 seconds
Batch generated in: 0.703682 seconds
Batch generated in: 0.689396 seconds
Batch generated in: 0.644848 seconds
Batch generated in: 0.636276 seconds
Total time: 3.308937 seconds
```